Updated webpack and babel and moved components into a folder.
