+++ package.json
@@ -7,18 +7,19 @@
   },
   "dependencies": {
     "animated-transition-group": "^1.0.15",
-    "history": "^4.6.1",
+    "history": "^4.6.3",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
     "react-redux": "^5.0.5",
-    "redux": "^3.6.0",
-    "redux-first-router": "1.4.6",
-    "redux-first-router-link": "^1.0.2",
-    "redux-first-router-restore-scroll": "1.1.1",
-    "redux-first-router-scroll-container": "1.0.1"
+    "redux": "^3.7.1",
+    "redux-first-router": "^1.4.18",
+    "redux-first-router-link": "^1.0.3",
+    "redux-first-router-restore-scroll": "^1.1.1",
+    "redux-first-router-scroll-container": "^1.0.1"
   },
   "devDependencies": {
     "babel-cli": "^6.24.1",
+    "babel-core": "^6.25.0",
     "babel-loader": "6.2.4",
     "babel-preset-es2015": "6.6.0",
     "babel-preset-react": "6.5.0",
@@ -26,8 +27,8 @@
     "css-loader": "0.23.1",
     "html-webpack-plugin": "1.6.1",
     "style-loader": "0.13.0",
-    "webpack": "1.12.13",
-    "webpack-dev-server": "1.14.1"
+    "webpack": "^3.0.0",
+    "webpack-dev-server": "^2.5.0"
   },
   "author": "WebpackBin",
   "license": "ISC"
