+++ webpack.config.js
@@ -22,7 +22,7 @@ module.exports = {
       {
         "test": /\.js?$/,
         "exclude": /node_modules/,
-        "loader": "babel",
+        "loader": "babel-loader",
         "query": {
           "presets": [
             "es2015",
@@ -34,8 +34,11 @@ module.exports = {
       },
       {
         "test": /\.css?$/,
-        "loader": "style!css?modules&localIdentName=[name]---[local]---[hash:base64:5]"
+        "loader": "style-loader!css-loader?modules&localIdentName=[name]---[local]---[hash:base64:5]"
       }
     ]
+  },
+  resolve: {
+    modules: ["node_modules", "src"]
   }
 }
